requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-huggingface
python-dotenv
openai
faiss-cpu
pandas
pypdf
